split: change convex trace to find splits on current edge
- add i parameter to cv_hull_trace_contour to allow the
  current edge to be specified in addition to the start
  and end edges. this is to allow search continuations.
- change convex trace to find splits on the current edge
  instead of the next edge, and update transform contour
  function to specify i parameter one past the start.
- hoist contour tracing debug heading and row into functions
- remove ab_s and change extract sign to return an int
  so that we can simply test winding direction against w
- remove i != s from winding direction tests because it is
  no longer necessary.
- rename split_at to split_idx, n_edges to n, start to s,
  end to e, scale to x

now that the convex trace finds splits on the current edge
instead of the next edge and the initial value of the i
parameter is provided, it is thus necessary for the i value
to be one past the start. edge start + dir (i=1 for forwards)
checks whether the second edge is non convex and the first
edge is tested at the end of the loop.

forward scan:

  cv_hull_trace_contour(ctx, idx, edges, n, 0, 1, n+1, 1, w);

reverse scan:

  cv_hull_trace_contour(ctx, idx, edges, n, n, n-1, -1, -1, -w)